requirements.txt
streamlit
faiss-cpu
pypdf
sentence-transformers
groq